+++ Urasandesu.Swathe/Urasandesu/Swathe/Metadata/BaseClass/BaseFieldMetadata.hpp
@@ -138,7 +138,7 @@ namespace Urasandesu { namespace Swathe { namespace Metadata { namespace BaseCla
     }
 
     template<class ApiHolder>    
-    ULONG BaseFieldMetadata<ApiHolder>::GetHashCode() const
+    size_t BaseFieldMetadata<ApiHolder>::GetHashCode() const
     {
         return Pimpl()->GetHashCode();
     }
